+++ src/rars/riscv/instructions/AND.java
@@ -34,7 +34,7 @@ public class AND extends Arithmetic {
                 "0000000", "111");
     }
 
-    public int compute(int value, int value2) {
+    public long compute(long value, long value2) {
         return value & value2;
     }
 }
