+++ src/rars/riscv/instructions/ORI.java
@@ -32,7 +32,7 @@ public class ORI extends ImmediateInstruction {
         super("ori t1,t2,-100", "Bitwise OR immediate : Set t1 to bitwise OR of t2 and sign-extended 12-bit immediate", "110");
     }
 
-    public int compute(int value, int immediate) {
+    public long compute(long value, long immediate) {
         return value | immediate;
     }
 }
