+++ src/rars/riscv/instructions/SB.java
@@ -35,8 +35,8 @@ public class SB extends Store {
         super("sb t1, -100(t2)", "Store byte : Store the low-order 8 bits of t1 into the effective memory byte address", "000");
     }
 
-    public void store(int address, int data) throws AddressErrorException {
-        Globals.memory.setByte(address, data & 0x000000FF);
+    public void store(int address, long data) throws AddressErrorException {
+        Globals.memory.setByte(address, (int)data & 0x000000FF);
     }
 }
