+++ src/rars/riscv/instructions/SLTIU.java
@@ -32,8 +32,7 @@ public class SLTIU extends ImmediateInstruction {
         super("sltiu t1,t2,-100", "Set less than immediate unsigned : If t2 is less than  sign-extended 16-bit immediate using unsigned comparison, then set t1 to 1 else set t1 to 0",
                 "011");
     }
-
-    public int compute(int value, int immediate) {
-        return (Integer.compareUnsigned(value, immediate) < 0) ? 1 : 0;
+    public long compute(long value, long immediate) {
+        return (Long.compareUnsigned(value, immediate) < 0) ? 1 : 0;
     }
 }
